+++ build_util/make_docs.py
@@ -1,12 +1,12 @@
 import json
 from pathlib import Path
 
-from voicevox_engine.dev.core import MockCoreWrapper
+from voicevox_engine.dev.core.mock import MockCoreWrapper
 from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
-from voicevox_engine.preset import PresetManager
-from voicevox_engine.setting import USER_SETTING_PATH, SettingLoader
+from voicevox_engine.preset.PresetManager import PresetManager
+from voicevox_engine.setting.SettingLoader import USER_SETTING_PATH, SettingLoader
 from voicevox_engine.tts_pipeline.tts_engine import CoreAdapter
-from voicevox_engine.utility import engine_root
+from voicevox_engine.utility.path_utility import engine_root
 
 
 def generate_api_docs_html(schema: str) -> str:
