+++ test/preset/test_preset.py
@@ -4,7 +4,9 @@ from shutil import copyfile
 from tempfile import TemporaryDirectory
 from unittest import TestCase
 
-from voicevox_engine.preset import Preset, PresetError, PresetManager
+from voicevox_engine.preset.Preset import Preset
+from voicevox_engine.preset.PresetError import PresetError
+from voicevox_engine.preset.PresetManager import PresetManager
 
 presets_test_1_yaml_path = Path("test/preset/presets-test-1.yaml")
 presets_test_2_yaml_path = Path("test/preset/presets-test-2.yaml")
