+++ test/test_mock_tts_engine.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from voicevox_engine.dev.tts_engine import MockTTSEngine
+from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
 from voicevox_engine.metas.Metas import StyleId
 from voicevox_engine.model import AccentPhrase, AudioQuery, Mora
 from voicevox_engine.tts_pipeline.kana_converter import create_kana
