--- voicevox_engine/dev/core/__init__.py
@@ -1,3 +1,0 @@
-from .mock import MockCoreWrapper
-
-__all__ = ["MockCoreWrapper"]
