--- voicevox_engine/dev/tts_engine/__init__.py
@@ -1,3 +1,0 @@
-from .mock import MockTTSEngine
-
-__all__ = ["MockTTSEngine"]
