--- voicevox_engine/engine_manifest/__init__.py
@@ -1,7 +1,0 @@
-from .EngineManifest import EngineManifest
-from .EngineManifestLoader import EngineManifestLoader
-
-__all__ = [
-    "EngineManifest",
-    "EngineManifestLoader",
-]
