--- voicevox_engine/preset/__init__.py
@@ -1,9 +1,0 @@
-from .Preset import Preset
-from .PresetError import PresetError
-from .PresetManager import PresetManager
-
-__all__ = [
-    "Preset",
-    "PresetManager",
-    "PresetError",
-]
