--- voicevox_engine/setting/__init__.py
@@ -1,9 +1,0 @@
-from .Setting import CorsPolicyMode, Setting
-from .SettingLoader import USER_SETTING_PATH, SettingLoader
-
-__all__ = [
-    "USER_SETTING_PATH",
-    "CorsPolicyMode",
-    "Setting",
-    "SettingLoader",
-]
