+++ voicevox_engine/tts_pipeline/tts_engine.py
@@ -572,7 +572,7 @@ def make_tts_engines_from_cores(cores: dict[str, CoreAdapter]) -> dict[str, TTSE
     tts_engines: dict[str, TTSEngine] = {}
     for ver, core in cores.items():
         if ver == MOCK_VER:
-            from ..dev.tts_engine import MockTTSEngine
+            from ..dev.tts_engine.mock import MockTTSEngine
 
             tts_engines[ver] = MockTTSEngine()
         else:
