+++ voicevox_engine/user_dict.py
@@ -12,7 +12,8 @@ from fastapi import HTTPException
 
 from .model import UserDictWord, WordTypes
 from .part_of_speech_data import MAX_PRIORITY, MIN_PRIORITY, part_of_speech_data
-from .utility import engine_root, get_save_dir, mutex_wrapper
+from .utility.mutex_utility import mutex_wrapper
+from .utility.path_utility import engine_root, get_save_dir
 
 root_dir = engine_root()
 save_dir = get_save_dir()
