--- voicevox_engine/utility/__init__.py
@@ -1,20 +1,0 @@
-from .connect_base64_waves import (
-    ConnectBase64WavesException,
-    connect_base64_waves,
-    decode_base64_waves,
-)
-from .core_version_utility import get_latest_core_version, parse_core_version
-from .mutex_utility import mutex_wrapper
-from .path_utility import delete_file, engine_root, get_save_dir
-
-__all__ = [
-    "ConnectBase64WavesException",
-    "connect_base64_waves",
-    "decode_base64_waves",
-    "get_latest_core_version",
-    "parse_core_version",
-    "delete_file",
-    "engine_root",
-    "get_save_dir",
-    "mutex_wrapper",
-]
